quit

+++ srcs/signals/signals.c
@@ -35,8 +35,11 @@ void	handle_sigquit(int sig)
 {
 	if (sig == SIGQUIT)
 	{
-		write(STDOUT_FILENO, "Quit (Core dumped)\n", 19);
-		kill(foreground_pid, SIGQUIT);
+		if (foreground_pid)
+		{
+			write(STDOUT_FILENO, "Quit (Core dumped)\n", 18);
+			kill(foreground_pid, SIGQUIT);
+		}
 	}
 }
